Reset text title after submit - Also refactored vote - Also added the appIcon to the app. Looks nice.
