+++ app.json
@@ -11,7 +11,7 @@
     ],
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./src/assets/icon.png",
+    "icon": "./src/assets/app_icon.png",
     "splash": {
       "image": "./src/assets/FAB_Logo.png",
       "resizeMode": "contain",
