+++ src/components/Vote/Vote.js
@@ -1,21 +1,19 @@
 import React, { useEffect, useState, useContext } from "react"
-import { View, Text,StyleSheet,Platform } from "react-native"
-import VoteScreen from "../../components/Vote/VoteScreen"
+import { View,StyleSheet,Platform } from "react-native"
+import VoteScreen from "./VoteScreen"
 import { SafeAreaView } from "react-native-safe-area-context"
 import fb from "../../db/init"
 import { getUserBadge } from "../../db/userBadge"
 import closeRoom from "../../db/closeRoom"
-import { StyledText } from "../../components/StyledText"
+import { StyledText } from "../StyledText"
 import { colors } from "../../constants/styles"
 import { getNumberOfVoters } from "../../db/Utility"
-import { AppContext} from "../../context/AppContext"
-import { CountDown } from "../../components/countdown/"
 import moment from "moment"
 import Swiper from 'react-native-deck-swiper';
 import Constants from 'expo-constants';
-import Button from "react-native-paper/src/components/Button";
-import {VoteContextProvider,VoteContext} from "../../components/Vote/VoteContext/VoteContext";
-const db = fb.database()
+import {VoteContext} from "./VoteContext/VoteContext";
+import Loader from "../FancyLoader/FancyLoader";
+const db = fb.database();
 
 
 const getTotalNumVoters = room => {
@@ -51,15 +49,16 @@ const roomStillActive = ({ room, roomID }) => {
     return false;
   }
   return true;
-}
+};
+
 
 const getActiveList = async () => {
-  const snapshot = await db.ref("rooms/active/").once("value")
-  let activeRooms = []
+  const snapshot = await db.ref("rooms/active/").once("value");
+  let activeRooms = [];
 
   for (var roomID of Object.keys(snapshot.val())) {
     if (roomStillActive({roomID:roomID, room: snapshot.val()[roomID]})) {
-      getTotalNumVoters(snapshot.val()[roomID])
+      getTotalNumVoters(snapshot.val()[roomID]);
 
       if (snapshot.val()[roomID]["meta_data"]["owner"] !== Constants.installationId &&
         noPreviousVote({ room: snapshot.val()[roomID], userID: Constants.installationId })) {
@@ -67,7 +66,7 @@ const getActiveList = async () => {
           numVotes: getTotalNumVoters(snapshot.val()[roomID]),
           id: roomID,
           room: snapshot.val()[roomID]
-        }
+        };
         activeRooms.push(currRoom)
       }
     }
@@ -76,10 +75,10 @@ const getActiveList = async () => {
   activeRooms.sort((a, b) => (a.numVotes > b.numVotes ? 1 : -1))
 
   return activeRooms
-}
+};
 
 const Vote = ({ navigation }) => {
-  const userID = Constants.installationId
+  const userID = Constants.installationId;
   //const [roomlist, setRoomList] = useState(null)
   const [badge, setBadge] = useState(null);
   const [hasSwipedAll, setHasSwipedAll] = useState(false);
@@ -88,75 +87,32 @@ const Vote = ({ navigation }) => {
 
   useEffect(() => {
     const getRooms = async () => {
-      const activeList = await getActiveList()
-      setRoomList(activeList)
+      const activeList = await getActiveList();
+      setRoomList(activeList);
       setCurrentRoom(0)
-    }
+    };
     getRooms()
-  }, [])
+  }, []);
 
   useEffect(() => {
     const getBadge = async () => {
       const dbbadge = await getUserBadge({ userID: userID })
       setBadge(dbbadge)
-    }
+    };
     getBadge()
-  }, [])
+  }, []);
 
   const handleNextRoom = () => {
     setCurrentRoom(currentRoom + 1)
-  }
-
-  /// for each room on roomlist, check if user is already on the voted list or is the owner of the post,
-  // if not then go through creating VoteScreens passing down the roomID
-  //once done prob have "voted on all" or something screen
-  /*{
-    "id": "room3",
-    "numVotes": 13,
-    "room": Object {
-      "meta_data": Object {
-        "owner": "jbrain98",
-        "time_created": "2020-02-24T00:22:20Z",
-        "title": "What watch is best for a first date?",
-      },
-      "optionA": Object {
-        "picture": "https://i.imgur.com/FQdEIKS.jpg",
-        "voters_influencer": Object {
-          "suzy": 1,
-        },
-        "voters_normal": Object {
-          "santa": 1,
-        },
-      },
-      "optionB": Object {
-        "picture": "https://i.imgur.com/7jZFeXM.jpg",
-        "voters_influencer": Object {
-          "zack": 1,
-        },
-        "voters_normal": Object {
-          "a": 1,
-          "b": 1,
-          "c": 1,
-          "d": 1,
-          "e": 1,
-          "f": 1,
-          "g": 1,
-          "h": 1,
-          "i": 1,
-          "suzy": 1,
-        },
-      },
-    },
-  }*/
-
-  const roomID = "room1"
-  //console.log(roomID)
+  };
 
   /*
   currently loading
   */
   if (!roomlist || !badge) {
-    return <StyledText>Loading..knm.</StyledText>
+    return <View style={{flex:1, alignItems:'center', justifyContent:'center'}}>
+            <Loader visible={true} />
+          </View>
   }
 
   /*
@@ -211,14 +167,6 @@ const Vote = ({ navigation }) => {
 }
 
 
-// So Vote has access to the context!!!
-const _Vote = ()=>(
-    <VoteContextProvider>
-      <Vote/>
-    </VoteContextProvider>
-);
-
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -240,4 +188,4 @@ const styles = StyleSheet.create({
 });
 
 
-export default _Vote
+export default Vote
