+++ src/views/index.js
@@ -1,6 +1,6 @@
 import Results from "./Results/Results"
 import Home from "./Home/Home"
 import Post from "./Post/Post"
-import {default as Vote} from "./Vote/Vote"
+import {default as Vote} from "./Vote/_Vote"
 import MyPostsPage from "./MyPostsPage/MyPostsPage"
 export { Results, Home, Post, Vote, MyPostsPage }
